refactor(core-run): split system-related into a single file
