+++ core/run/src/components/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod system;
+
 #[cfg(all(
     not(target_env = "msvc"),
     not(target_os = "macos"),
