+++ core/run/src/lib.rs
@@ -1,8 +1,6 @@
 #![allow(clippy::uninlined_format_args, clippy::mutable_key_type)]
 
-use std::{collections::HashMap, panic::PanicInfo, path::Path, sync::Arc, time::Duration};
-
-use backtrace::Backtrace;
+use std::{collections::HashMap, path::Path, sync::Arc, time::Duration};
 
 use common_apm::metrics::mempool::{MEMPOOL_CO_QUEUE_LEN, MEMPOOL_LEN_GAUGE};
 use common_apm::{server::run_prometheus_server, tracing::global_tracer_register};
@@ -11,8 +9,6 @@ use common_config_parser::types::{Config, ConfigJaeger, ConfigPrometheus, Config
 use common_crypto::{BlsPrivateKey, BlsPublicKey, Secp256k1, Secp256k1PrivateKey, ToPublicKey};
 
 use protocol::codec::{hex_decode, ProtocolCodec};
-#[cfg(unix)]
-use protocol::tokio::signal::unix as os_impl;
 use protocol::tokio::{
     self, runtime::Builder as RuntimeBuilder, sync::Mutex as AsyncMutex, time::sleep,
 };
@@ -370,8 +366,7 @@ impl Axon {
         // Run consensus
         Self::run_overlord_consensus(metadata, validators, current_block, overlord_consensus);
 
-        Self::set_ctrl_c_handle().await;
-
+        components::system::set_ctrl_c_handle().await;
         components::profiling::stop();
 
         Ok(())
@@ -713,58 +708,6 @@ impl Axon {
                 log::error!("axon-consensus: {:?} error", e);
             }
         });
-    }
-
-    async fn set_ctrl_c_handle() {
-        let ctrl_c_handler = tokio::spawn(async {
-            #[cfg(windows)]
-            let _ = tokio::signal::ctrl_c().await;
-            #[cfg(unix)]
-            {
-                let mut sigtun_int = os_impl::signal(os_impl::SignalKind::interrupt()).unwrap();
-                let mut sigtun_term = os_impl::signal(os_impl::SignalKind::terminate()).unwrap();
-                tokio::select! {
-                    _ = sigtun_int.recv() => {}
-                    _ = sigtun_term.recv() => {}
-                };
-            }
-        });
-
-        // register channel of panic
-        let (panic_sender, mut panic_receiver) = tokio::sync::mpsc::channel::<()>(1);
-
-        std::panic::set_hook(Box::new(move |info: &PanicInfo| {
-            let panic_sender = panic_sender.clone();
-            Self::panic_log(info);
-            panic_sender.try_send(()).expect("panic_receiver is droped");
-        }));
-
-        tokio::select! {
-            _ = ctrl_c_handler => { log::info!("ctrl + c is pressed, quit.") },
-            _ = panic_receiver.recv() => { log::info!("child thread panic, quit.") },
-        };
-    }
-
-    fn panic_log(info: &PanicInfo) {
-        let backtrace = Backtrace::new();
-        let thread = std::thread::current();
-        let name = thread.name().unwrap_or("unnamed");
-        let location = info.location().unwrap(); // The current implementation always returns Some
-        let msg = match info.payload().downcast_ref::<&'static str>() {
-            Some(s) => *s,
-            None => match info.payload().downcast_ref::<String>() {
-                Some(s) => &**s,
-                None => "Box<Any>",
-            },
-        };
-        log::error!(
-            target: "panic", "thread '{}' panicked at '{}': {}:{} {:?}",
-            name,
-            msg,
-            location.file(),
-            location.line(),
-            backtrace,
-        );
     }
 }
